requirements.txt
streamlit
requests
pymupdf
pdf2image
pytesseract
Pillow
python-dotenv
google-generativeai
ollama